Keep the Dag processor's tests in the file that mirrors its module
Airflow's only stated rule for test layout is that a test file mirrors
its source module, so anything covering dag_processing/manager.py
belongs in test_manager.py. Splitting by topic instead put the sweep
tests out of reach of the helpers and fixtures the class already
provided, and the copies written to replace them were free to drift.
